backend/requirements.txt
# FastAPI core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
aiosqlite==0.20.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation & Serialization
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2

# Pagination
sqlalchemy-utils==0.41.2
